+++ frontend/src/components/InstanceForm/Buttons.vue
@@ -77,12 +77,16 @@ import {
 } from "./common";
 import { useInstanceFormContext } from "./context";
 
-withDefaults(
+const props = withDefaults(
   defineProps<{
     allowCancel?: boolean;
+    onCreated?: (instance: Instance) => void;
+    onUpdated?: (instance: Instance) => void;
   }>(),
   {
     allowCancel: true,
+    onCreated: undefined,
+    onUpdated: undefined,
   }
 );
 
@@ -280,7 +284,14 @@ const doCreate = async () => {
       useDatabaseV1Store().searchDatabases({
         filter: `instance = "${createdInstance.name}"`,
       });
-      router.push(`/${createdInstance.name}`);
+
+      if (props.onCreated) {
+        props.onCreated(createdInstance);
+      } else {
+        router.push(`/${createdInstance.name}`);
+        events.emit("dismiss");
+      }
+
       pushNotification({
         module: "bytebase",
         style: "SUCCESS",
@@ -292,7 +303,6 @@ const doCreate = async () => {
     });
   } finally {
     state.value.isRequesting = false;
-    events.emit("dismiss");
   }
 };
 
@@ -493,6 +503,10 @@ const doUpdate = async () => {
         updatedInstance.title,
       ]),
     });
+
+    if (props.onUpdated) {
+      props.onUpdated(updatedInstance);
+    }
   } finally {
     state.value.isRequesting = false;
   }
