+++ frontend/src/components/v2/Model/Instance/InstanceV1Table/InstanceV1Table.vue
@@ -94,6 +94,7 @@ const columnList = computed((): InstanceDataTableColumn[] => {
   const ENVIRONMENT: InstanceDataTableColumn = {
     key: "environment",
     title: t("common.environment"),
+    className: 'whitespace-nowrap',
     resizable: true,
     render: (instance) => (
       <EnvironmentV1Name
