+++ frontend/src/views/sql-editor/AsidePanel/GutterBar/GutterBar.vue
@@ -21,14 +21,14 @@
     <OpenAIButton :size="size" />
 
     <HideInStandaloneMode>
-      <SettingButton :size="size" />
+      <SettingButton :style="buttonStyle" v-bind="buttonProps" />
     </HideInStandaloneMode>
   </div>
 </template>
 
 <script setup lang="ts">
 import { storeToRefs } from "pinia";
-import { computed, watch } from "vue";
+import { computed, toRef, watch } from "vue";
 import HideInStandaloneMode from "@/components/misc/HideInStandaloneMode.vue";
 import {
   useConnectionOfCurrentSQLEditorTab,
@@ -38,12 +38,12 @@ import {
 import { UNKNOWN_ID } from "@/types";
 import { hasProjectPermissionV2, instanceV1HasAlterSchema } from "@/utils";
 import { useSQLEditorContext, type AsidePanelTab } from "../../context";
+import { SettingButton } from "../../Setting";
 import OpenAIButton from "./OpenAIButton.vue";
-import SettingButton from "./SettingButton.vue";
 import TabItem from "./TabItem.vue";
-import type { Size } from "./common";
+import { useButton, type Size } from "./common";
 
-withDefaults(
+const props = withDefaults(
   defineProps<{
     size?: Size;
   }>(),
@@ -56,6 +56,12 @@ const me = useCurrentUserV1();
 const { currentTab, isDisconnected } = storeToRefs(useSQLEditorTabStore());
 const { asidePanelTab } = useSQLEditorContext();
 const { instance, database } = useConnectionOfCurrentSQLEditorTab();
+
+const { props: buttonProps, style: buttonStyle } = useButton({
+  size: toRef(props, "size"),
+  active: false,
+  disabled: false,
+});
 
 const isSchemalessInstance = computed(() => {
   if (instance.value.uid === String(UNKNOWN_ID)) {
