+++ frontend/src/views/sql-editor/Setting/Sidebar/index.ts
@@ -1,3 +1,4 @@
 import Sidebar from "./Sidebar.vue";
 
+export * from "./common";
 export default Sidebar;
