+++ frontend/src/views/sql-editor/Setting/index.ts
@@ -1,4 +1,5 @@
+import SettingButton from "./SettingButton.vue";
 import Sidebar from "./Sidebar";
 
 export * from "./context";
-export { Sidebar };
+export { Sidebar, SettingButton };
